remove debugging bits, add extra not_if to awscli install

+++ identity_base_config/recipes/default.rb
@@ -11,6 +11,7 @@ end
 
 execute 'install awscli as needed' do
   command 'pip3 install awscli'
+  not_if { ::File.exist?('/usr/bin/aws') }
   not_if { ::File.exist?('/usr/local/bin/aws') }
 end
