+++ static_eip/resources/assign.rb
@@ -24,6 +24,7 @@ action :create do
 
   execute 'install awscli as needed' do
     command 'pip3 install awscli'
+    not_if { ::File.exist?('/usr/bin/aws') }
     not_if { ::File.exist?('/usr/local/bin/aws') }
   end
